Armino SDK 1.3.10 1. 丰富多媒体功能及接口 2. 优化文档及部分示例代码

+++ middleware/driver/include/bk_private/bk_usb.h
@@ -125,9 +125,9 @@ typedef enum
 
 typedef struct
 {
-    uint32_t  fps;
-    uint32_t  width;
-    uint32_t  height;
+    uint16_t  fps;
+    uint16_t  width;
+    uint16_t  height;
 } UVC_ResolutionFramerate;
 
 /*
